chore: added WalletDataProvider.tsx for global test
